pyproject.toml
[project]
name = "anpr"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "cycler>=0.12.1",
    "easyocr>=1.7.2",
    "fastapi>=0.115.12",
    "joblib>=1.5.0",
    "kiwisolver>=1.4.8",
    "matplotlib>=3.10.3",
    "numpy>=2.2.6",
    "opencv-python>=4.11.0.86",
    "opencv-python-headless>=4.11.0.86",
    "pillow>=11.2.1",
    "pyparsing>=3.2.3",
    "pytesseract>=0.3.13",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.1.0",
    "python-multipart>=0.0.20",
    "requests>=2.32.3",
    "scikit-learn>=1.6.1",
    "scipy>=1.15.3",
    "six>=1.17.0",
    "sqlalchemy>=2.0.41",
    "torch>=2.7.0",
    "torchaudio>=2.7.0",
    "torchvision>=0.22.0",
    "uvicorn>=0.34.2",
    "wincertstore>=0.2.1",
]
